+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix `buf convert` to allow for zero length for `binpb`, `txtpb`, and `yaml` formats.
 - Fix use of deprecated flag `--include-types` for `buf generate`.
+- Fix type filter with unused image dependencies for `buf generate`.
 
 ## [v1.50.1] - 2025-03-10